PhoneSignIn renamed as Login.vue
